created extensions methods for endpoints
