+++ RangoAgil/RangoAgil.API/Program.cs
@@ -1,11 +1,6 @@
-using AutoMapper;
-using Microsoft.AspNetCore.Http.HttpResults;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using RangoAgil.API.DbContexts;
-using RangoAgil.API.EndPoitHandlers;
-using RangoAgil.API.Entities;
-using RangoAgil.API.Models;
+using RangoAgil.API.Extensions;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -60,24 +55,7 @@ app.MapGet("/", () => "Bem vindo ao Rango's");
 
 #endregion
 
-var rangosEndpoints = app.MapGroup("/rangos");
-var rangosComIdEndpoints = rangosEndpoints.MapGroup("/{rangoId:int}");
-var rangosComIngredientesEndpoints = rangosComIdEndpoints.MapGroup("/ingredientes");
-
-//using automapper
-
-rangosEndpoints.MapGet("", RangosHandlers.GetRangosAsync);
-
-rangosEndpoints.MapPost("", RangosHandlers.CreateRangoAsync);
-
-rangosComIdEndpoints.MapGet("", RangosHandlers.GetRangoById)
-                    .WithName("GetRangoById");
-
-rangosComIdEndpoints.MapPut("", RangosHandlers.UpdateRangoAsync);
-
-rangosComIdEndpoints.MapDelete("", RangosHandlers.DeleteRangoAsync);
-
-rangosComIngredientesEndpoints.MapGet("", IngredientesHandlers.GetIngredientesByRangoIdAsync);
-
+app.RegisterRangosEndpoints();
+app.RegisterIngredientesEndpoints();
 
 app.Run();
